move theme files to appropriate packages
