+++ presentation/designSystem/src/main/kotlin/com/berlin/aflami/ui/theme/Theme.kt
@@ -1,11 +1,11 @@
-package com.berlin.aflami.ui.theme.theme
+package com.berlin.aflami.ui.theme
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ReadOnlyComposable
-import com.berlin.aflami.ui.theme.textstyle.AflamiTextStyle
-import com.berlin.aflami.ui.theme.textstyle.LocalAflamiTextStyle
-import com.berlin.aflami.ui.theme.color.AflamiColors
-import com.berlin.aflami.ui.theme.color.LocalAflamiColors
+import com.berlin.aflami.ui.textstyle.AflamiTextStyle
+import com.berlin.aflami.ui.textstyle.LocalAflamiTextStyle
+import com.berlin.aflami.ui.color.AflamiColors
+import com.berlin.aflami.ui.color.LocalAflamiColors
 
 object Theme {
     val color: AflamiColors
